requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
aiohttp==3.10.11
lxml==5.3.0
python-multipart==0.0.19
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
stripe==10.12.0
python-dotenv==1.0.1
